backend/requirements.txt.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
sentence-transformers
chromadb
pydantic
python-dotenv
sqlalchemy
unstructured
PyPDF2
python-multipart
